# Used numpy for much faster python implementation