--- a/run.sh
+++ b/run.sh
@@ -52,11 +52,11 @@ run_rust_rayon() {
         fi
 }
 
-run_python() {
+run_python_np() {
     echo "Running Python with Numpy" &&
         cd ./python &&
         if [ $HYPER == 1 ]; then
-            command hyperfine -r 1 "python3 ./related_np.py"
+            command hyperfine -r 5 --show-output "python3 ./related_np.py"
         else
             command time -f '%es %Mk' python3 ./related_np.py
         fi
@@ -98,10 +98,15 @@ elif [ "$first_arg" = "rust_ray" ]; then
     run_rust_rayon &&
         check_output "related_posts_rust_rayon.json"
 
-elif [ "$first_arg" = "python" ]; then
+elif [ "$first_arg" = "py" ]; then
 
     run_python &&
         check_output "related_posts_python.json"
+
+elif [ "$first_arg" = "numpy" ]; then
+
+    run_python_np &&
+        check_output "related_posts_python_np.json"
 
 elif [ "$first_arg" = "all" ]; then
 
@@ -124,5 +129,5 @@ elif [ "$first_arg" = "clean" ]; then
         cd rust_rayon && cargo clean
 
 else
-    echo "Valid args: go | rust | python | all | clean. Unknown argument: $first_arg"
+    echo "Valid args: go | rust | py | numpy | all | clean. Unknown argument: $first_arg"
 fi