This numpy implementation runs in ~0.72sec on my m2 air.

Edit: Knocked another 100ms off with array broadcasting, main now runs in 0.65 sec, 0.54sec without IO.