@jinyus 2023-09-24
Thanks for the PR. I got roughly the same speed on my machine